refactor: reorganize types by function
