+++ src/validators/index.ts
@@ -12,3 +12,4 @@ export { validateScripts } from "./validateScripts.ts";
 export { validateType } from "./validateType.ts";
 export { validateUrlOrMailto } from "./validateUrlOrMailto.ts";
 export { validateVersion } from "./validateVersion.ts";
+export type { People, Person, UrlType } from "./validation.types.ts";
